chore(test): standard test ext name; better organize utils #50
