--- src/package/SlateEditor/__tests__/SlateEditor.test.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import { mount } from 'enzyme'
-import SlateEditor from '../SlateEditor'
-
-describe('SlateEditor', () => {
-  it('renders without crashing', () => {
-    mount(<SlateEditor />)
-  })
-})
