+++ src/package/slate-editor/SlateContent.js
