+++ src/package/slate-editor/SlateToolbar.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import classnames from 'classnames'
-import {cloneElement} from './Utils'
+import { cloneElement } from '../utils/react'
 
 export default ({ children, style, className, ...rest }) => (
   <div className={classnames('editor--toolbar', className)} style={style}>
