+++ src/package/slate-editor/index.js
