+++ src/package/slate-editor/initialEditorState.js
