+++ src/package/utils/react/index.js
@@ -1,0 +1,1 @@
+export { default as cloneElement } from './clone-element'
