+++ src/package/utils/type-check.js
@@ -1,0 +1,1 @@
+export const isFunction = obj => !!(obj && obj.constructor === Function)
